+++ CHANGELOG.md
@@ -3,9 +3,17 @@
 0.19.0 - (2022-02-XX)
 ------------------
 
+A few more changes to projection file blending:
 * Added a few more arguments to `music_into_models.py` to be able to make selecting the production
 list of models easier. Now input args can be saved to file, and lists of models can be passed in
 as json.
+* Switched phash computation to only take place on the bounding box regions of the images. 
+This results to tighter looking overlays.
+* Can use `--wav` argument multiple times now, good for working through projection files with
+multiple tracks.
+* Audio files are now scaled in duration to the length of the video contained in the projection file.
+This way, the number of FFTs, and the number of final latents to blend are always the same. There's
+a check to make sure the projected frame count and the # of vectors are not too far off. 
 
 
 0.18.0 - (2022-02-08)